Added updated navigation to partner profile details.
