+++ frontend/src/components/partners/profile/partnerProfileHeader.js
@@ -1,28 +1,43 @@
-
-import PropTypes from 'prop-types';
-import IconButton from 'material-ui/IconButton';
-import React from 'react';
-import MoreVert from 'material-ui-icons/MoreVert';
-import Cancel from 'material-ui-icons/Cancel';
-import Menu, { MenuItem } from 'material-ui/Menu';
+import React, { Component } from 'react';
+import { connect } from 'react-redux';
+import { browserHistory as history } from 'react-router';
 import { withStyles, createStyleSheet } from 'material-ui/styles';
+import VerifiedUser from 'material-ui-icons/VerifiedUser';
+import Flag from 'material-ui-icons/Flag';
+import PropTypes from 'prop-types';
+import Typography from 'material-ui/Typography';
+import PartnerProfileHeaderMenu from './partnerProfileHeaderMenu';
+import HeaderNavigation from '../../../components/common/headerNavigation';
 
-const messages = {
-  deactivate: 'Deactivate Profile',
-};
 
-const styleSheet = createStyleSheet('HqProfileOverviewHeader', (theme) => {
-  const paddingSmall = theme.spacing.unit * 3;
+const styleSheet = createStyleSheet('CountryOfficesHeader', (theme) => {
   const paddingIcon = theme.spacing.unit;
 
   return {
-    root: {
+    alignCenter: {
       display: 'flex',
       alignItems: 'center',
-      justifyContent: 'flex-end',
-      paddingRight: `${paddingSmall}px`,
     },
-    icon: {
+    iconNotVerified: {
+      fill: theme.palette.primary[500],
+      width: 20,
+      height: 20,
+      margin: `0 0 0 ${paddingIcon}px`,
+    },
+    iconVerified: {
+      fill: '#009A54',
+      width: 20,
+      height: 20,
+      margin: `0 0 0 ${paddingIcon}px`,
+    },
+    iconYellow: {
+      fill: '#FFC400',
+      width: 20,
+      height: 20,
+      margin: `0 ${paddingIcon}px 0 0`,
+    },
+    iconRed: {
+      fill: '#D50000',
       width: 20,
       height: 20,
       margin: `0 ${paddingIcon}px 0 0`,
@@ -30,52 +45,80 @@ const styleSheet = createStyleSheet('HqProfileOverviewHeader', (theme) => {
   };
 });
 
+const PartnerTitle = (props) => {
+  const { classes, partner } = props;
 
-class PartnerProfileHeader extends React.Component {
+  return (
+    <div className={classes.alignCenter}>
+      <Typography type="headline">
+        {partner.name}
+      </Typography>
+      <VerifiedUser className={partner.verified ? classes.iconVerified : classes.iconNotVerified} />
+      {partner.flagYellow && <Flag className={classes.iconYellow} />}
+      {partner.flagRed && <Flag className={classes.iconRed} />}
+    </div>);
+};
+
+class PartnerProfileHeader extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      anchorPoint: undefined,
-      open: false,
+      index: 0,
     };
-
-    this.handleOpenMenu = this.handleOpenMenu.bind(this);
-    this.handleRequestClose = this.handleRequestClose.bind(this);
+    this.handleChange = this.handleChange.bind(this);
   }
 
-  handleClick(event) {
-    this.setState({ open: true, anchorPoint: event.currentTarget });
+  partnerTitle() {
+    const { classes, partner } = this.props;
+
+    return (
+      <div className={classes.alignCenter}>
+        <Typography type="headline">
+          {partner.name}
+        </Typography>
+        <VerifiedUser
+          className={partner.verified ? classes.iconVerified : classes.iconNotVerified}
+        />
+        {partner.flagYellow && <Flag className={classes.iconYellow} />}
+        {partner.flagRed && <Flag className={classes.iconRed} />}
+      </div>);
   }
 
-  handleOpenMenu() {
-    this.setState({ open: false });
+  updatePath() {
+    const { tabs, location, partnerId } = this.props;
+    if (tabs.findIndex(tab => location.match(`^/partner/${partnerId}/${tab.path}`)) === -1) {
+      history.push('/');
+    }
+
+    return tabs.findIndex(tab => location.match(`^/partner/${partnerId}/${tab.path}`));
   }
 
-  handleRequestClose() {
-    this.setState({ open: false });
+  handleChange(event, index) {
+    const { tabs, partnerId } = this.props;
+    history.push(`/partner/${partnerId}/${tabs[index].path}`);
   }
 
   render() {
-    const { classes } = this.props;
+    const {
+      tabs,
+      children,
+    } = this.props;
+
+    const index = this.updatePath();
 
     return (
-      <div className={classes.root}>
-        <IconButton onClick={event => this.handleClick(event)} >
-          <MoreVert />
-        </IconButton>
-        <Menu
-          id="deactivateMenu"
-          anchorEl={this.state.anchorPoint}
-          open={this.state.open}
-          onRequestClose={this.handleRequestClose}
+      <div>
+        <HeaderNavigation
+          backButton
+          tabs={tabs}
+          index={index}
+          handleBackButton={() => { history.goBack(); }}
+          header={<PartnerProfileHeaderMenu handleMoreClick={() => { }} />}
+          titleObject={PartnerTitle(this.props)}
+          handleChange={this.handleChange}
         >
-          <MenuItem onClick={this.handleRequestClose}>
-            <div className={classes.root}>
-              <Cancel className={classes.icon} />
-              {messages.deactivate}
-            </div>
-          </MenuItem>
-        </Menu>
+          {children}
+        </HeaderNavigation>
       </div>
     );
   }
@@ -83,7 +126,35 @@ class PartnerProfileHeader extends React.Component {
 
 PartnerProfileHeader.propTypes = {
   classes: PropTypes.object.isRequired,
-  handleMoreClick: PropTypes.func.isRequired,
+  tabs: PropTypes.array.isRequired,
+  children: PropTypes.node,
+  location: PropTypes.string.isRequired,
+  partner: PropTypes.object.isRequired,
+  partnerId: PropTypes.string.isRequired,
 };
 
-export default withStyles(styleSheet)(PartnerProfileHeader);
+PartnerTitle.propTypes = {
+  classes: PropTypes.object.isRequired,
+  partner: PropTypes.object.isRequired,
+};
+
+const mapStateToProps = (state, ownProps) => ({
+  tabs: state.agencyPartnerProfileNav.tabs,
+  partner: state.agencyPartnerProfile[ownProps.params.id],
+  profile: state.hqProfile[ownProps.params.id],
+  partnerId: ownProps.params.id,
+  location: ownProps.location.pathname,
+});
+
+const mapDispatchToProps = () => ({
+  onItemClick: (id, path) => {
+    history.push(path);
+  },
+});
+
+const connectedPartnerProfile = connect(mapStateToProps, mapDispatchToProps)(PartnerProfileHeader);
+const connectedPartnerTitle = connect(mapStateToProps, mapDispatchToProps)(PartnerTitle);
+withStyles(styleSheet)(connectedPartnerTitle);
+
+export default withStyles(styleSheet)(connectedPartnerProfile);
+
