+++ frontend/src/reducers/agencyPartnerProfileNav.js
@@ -3,10 +3,10 @@ import { SESSION_INIT } from './session';
 
 const initialState = {
   tabs: [
-    { path: '/overview', label: 'Overview', name: 'a' },
-    { path: '/details', label: 'Profile details', name: 'a' },
-    { path: '/partner/info/users', label: 'Users', name: 'a' },
-    { path: '/partner/info/applications', label: 'Applications', name: 'a' },
+    { path: 'overview', label: 'Overview', name: 'a' },
+    { path: 'details', label: 'Profile details', name: 'a' },
+    { path: 'users', label: 'Users', name: 'a' },
+    { path: 'applications', label: 'Applications', name: 'a' },
   ],
 };
